Adding a half working camera

+++ Viewer/include/Renderer.h
@@ -40,6 +40,6 @@ public:
 	void SwapBuffers();
 	void ClearColorBuffer(const glm::vec3& color);
 	void SetViewport(int viewportWidth, int viewportHeight, int viewportX = 0, int viewportY = 0);
-	void drawLine(int x1, int y1, int x2, int y2, glm::vec4 color = glm::vec4(0.0f, 0.0f, 0.0f, 1.0f));
+	void drawLine(float x1, float y1, float x2, float y2, glm::vec4 color = glm::vec4(0.0f, 0.0f, 0.0f, 1.0f));
 	// Add more methods/functionality as needed...
 };
